# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KNXVault API
  version: 0.4.7
  description: Secrets management and PKI REST API (Phase 1 MVP + Phase 2 complete).
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8200
    description: Local development
...
paths:
...
  /sys/rotate-master-key:
    post:
      tags: [sys]
      summary: Rotate envelope master key
      operationId: rotateMasterKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotateMasterKeyRequest"
      responses:
        "200":
          description: Master key rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateMasterKeyResponse"
...
  /sys/generate-unseal-shares:
    post:
      tags: [sys]
      summary: Split unseal key into Shamir shares
      operationId: generateUnsealShares
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, shares, threshold]
              properties:
                key:
                  type: string
                  description: Base64 full unseal key
                shares:
                  type: integer
                  minimum: 2
                threshold:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Base64 Shamir shares
  /sys/raft/add-node:
    post:
      tags: [sys]
      summary: Add Raft cluster member
      operationId: raftAddNode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaftAddNodeRequest"
      responses:
        "200":
          description: Node added
  /sys/raft/remove-node:
    post:
      tags: [sys]
      summary: Remove Raft cluster member
      operationId: raftRemoveNode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaftRemoveNodeRequest"
      responses:
        "200":
          description: Node removed
...
  /pki/root:
    post:
      tags: [pki]
      summary: Create a root CA
      operationId: createRootCA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRootCARequest"
      responses:
        "201":
          description: Root CA created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CAResponse"
...
  /pki/issue:
    post:
      tags: [pki]
      summary: Issue a leaf certificate
      operationId: issueCertificate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCertRequest"
      responses:
        "201":
          description: Certificate issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCertResponse"
...
  /sys/policies/{name}:
    put:
      tags: [sys]
      summary: Create or update RBAC policy
      operationId: putPolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyRequest"
      responses:
        "204": { description: Policy saved }
    get:
      tags: [sys]
      summary: Get RBAC policy
      operationId: getPolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
    delete:
      tags: [sys]
      summary: Delete RBAC policy
      operationId: deletePolicy
      security: [{ bearerAuth: [] }]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Policy deleted }
...
  /sys/exposure/report:
    post:
      tags: [sys]
      summary: Report secret exposure (HMAC + timestamp signed)
      operationId: reportExposure
      parameters:
        - name: X-KNXVault-Exposure-Signature
          in: header
          required: true
          schema: { type: string }
        - name: X-KNXVault-Exposure-Timestamp
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Report accepted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    RotateMasterKeyRequest:
      type: object
      required: [new_key]
      properties:
        new_key:
          type: string
          description: Base64-encoded 32-byte master key
    RotateMasterKeyResponse:
      type: object
      properties:
        key_version:
          type: integer
...
    RaftAddNodeRequest:
      type: object
      required: [node_id, address]
      properties:
        node_id:
          type: integer
          format: int64
        address:
          type: string
    RaftRemoveNodeRequest:
      type: object
      required: [node_id]
      properties:
        node_id:
          type: integer
          format: int64
...
    CreateRootCARequest:
      type: object
      required: [name, common_name, ttl]
      properties:
        name:
          type: string
        common_name:
          type: string
        ttl:
          type: string
        key_type:
          type: string
        key_bits:
          type: integer
...
    IssueCertRequest:
      type: object
      required: [role, common_name]
      properties:
        role:
          type: string
        common_name:
          type: string
        dns_names:
          type: array
          items:
            type: string
        ip_addresses:
          type: array
          items:
            type: string
        ttl:
          type: string
        key_bits:
          type: integer
        auto_renew:
          type: boolean
          description: Track and auto-renew this certificate before expiry
...
    CAResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        cert_pem:
          type: string
        serial:
          type: string
        expires_at:
          type: string
    IssueCertResponse:
      type: object
      properties:
        cert_pem:
          type: string
        private_key_pem:
          type: string
        serial:
          type: string
        expires_at:
          type: string
...
    PolicyRequest:
      type: object
      required: [effect, resources, actions]
      properties:
        effect: { type: string }
        resources:
          type: array
          items: { type: string }
        actions:
          type: array
          items: { type: string }
        conditions:
          type: object
          additionalProperties: true
    PolicyResponse:
      allOf:
        - $ref: "#/components/schemas/PolicyRequest"
        - type: object
          properties:
            name: { type: string }
...