# internal/api/handlers/sys.go
// Copyright Kubenexis Systems Private Limited.
// SPDX-License-Identifier: Apache-2.0

package handlers

import (
	"context"
	"crypto/sha256"
	"encoding/base64"
	"encoding/hex"
	"fmt"
	"net"
	"net/http"
	"os"
	"path/filepath"
	"strings"
	"time"

	"github.com/gin-gonic/gin"

	"github.com/kubenexis/knxvault/internal/api/dto"
	"github.com/kubenexis/knxvault/internal/auth"
	"github.com/kubenexis/knxvault/internal/crypto/shamir"
	"github.com/kubenexis/knxvault/internal/domain/common"
	pkiengine "github.com/kubenexis/knxvault/internal/engine/pki"
	"github.com/kubenexis/knxvault/internal/netutil"
	"github.com/kubenexis/knxvault/internal/notify"
	"github.com/kubenexis/knxvault/internal/service"
	"github.com/kubenexis/knxvault/internal/sys"
	"github.com/kubenexis/knxvault/internal/utils"
)

// SealController controls operational seal state.
type SealController interface {
	Sealed() bool
	Seal()
	Unseal(key []byte) bool
}

// MultiShareUnsealer is optional Shamir share support on the seal controller.
type MultiShareUnsealer interface {
	SubmitShare(share []byte) (unsealed bool, have, need int, errMsg string)
	UnsealProgress() (have, need int)
}

// RaftMembership controls cluster membership changes.
type RaftMembership interface {
	AddNode(ctx context.Context, nodeID uint64, address string) error
	RemoveNode(ctx context.Context, nodeID uint64) error
	IsLeader() bool
}

// SysHandler serves system endpoints.
type SysHandler struct {
	auth            *auth.Service
	pki             *service.PKIService
	database        *service.DatabaseService
	rotation        *service.RotationService
	orchestration   *service.OrchestrationService
	masterKey       *service.MasterKeyService
	seal            SealController
	raft            RaftMembership
	masterKeyBytes  []byte
	exposureAuto    bool
	exposureWebhook *notify.Webhook
	unsealAllowNets []*net.IPNet
}

// SetUnsealAllowNets configures source IP allowlist for unseal (W75-04).
func (h *SysHandler) SetUnsealAllowNets(nets []*net.IPNet) {
	if h != nil {
		h.unsealAllowNets = nets
	}
}

// NewSysHandler constructs a SysHandler.
func NewSysHandler(
	svc *auth.Service,
	pki *service.PKIService,
	database *service.DatabaseService,
	rotation *service.RotationService,
	orchestration *service.OrchestrationService,
	masterKey *service.MasterKeyService,
	seal SealController,
	raft RaftMembership,
	masterKeyBytes []byte,
	exposureAuto bool,
	exposureWebhook *notify.Webhook,
) *SysHandler {
	return &SysHandler{
		auth:            svc,
		pki:             pki,
		database:        database,
		rotation:        rotation,
		orchestration:   orchestration,
		masterKey:       masterKey,
		seal:            seal,
		raft:            raft,
		masterKeyBytes:  masterKeyBytes,
		exposureAuto:    exposureAuto,
		exposureWebhook: exposureWebhook,
	}
}

// Capabilities handles GET /sys/capabilities.
func (h *SysHandler) Capabilities(c *gin.Context) {
	principal, ok := auth.PrincipalFromContext(c.Request.Context())
	if !ok || h.auth == nil {
		c.JSON(http.StatusOK, dto.CapabilitiesResponse{Capabilities: []string{}})
		return
	}

	c.JSON(http.StatusOK, dto.CapabilitiesResponse{
		Capabilities: h.auth.Capabilities(principal),
	})
}

// Init handles POST /sys/init one-time bootstrap.
func (h *SysHandler) Init(c *gin.Context) {
	var req dto.InitRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	fp := ""
	if len(h.masterKeyBytes) > 0 {
		sum := sha256.Sum256(h.masterKeyBytes)
		fp = hex.EncodeToString(sum[:8])
	}
	if err := sys.MarkInitialized(fp); err != nil {
		_ = c.Error(common.Wrap(common.ErrCodeValidation, err.Error(), err))
		return
	}
	var caResp *dto.CAResponse
	if req.CreateRootCA && h.pki != nil {
		name := req.RootCAName
		if name == "" {
			name = "root"
		}
		cn := req.RootCommonName
		if cn == "" {
			cn = "KNXVault Root CA"
		}
		result, err := h.pki.CreateRoot(c.Request.Context(), pkiengine.CreateRootRequest{
			Name:       name,
			CommonName: cn,
			TTL:        "87600h",
		})
		if err != nil {
			_ = c.Error(err)
			return
		}
		caResp = &dto.CAResponse{
			ID:        result.ID.String(),
			Name:      result.Name,
			CertPEM:   result.CertPEM,
			Serial:    result.Serial,
			ExpiresAt: result.ExpiresAt.Format(time.RFC3339),
		}
	}
	c.JSON(http.StatusOK, gin.H{
		"initialized":            true,
		"master_key_fingerprint": fp,
		"root_ca":                caResp,
	})
}

// ReportExposure handles POST /sys/exposure/report.
func (h *SysHandler) ReportExposure(c *gin.Context) {
	var req dto.ExposureReportRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	actions := make([]string, 0, 2)
	if h.exposureAuto {
		if req.LeaseID != "" && h.database != nil {
			if _, err := h.database.Revoke(c.Request.Context(), req.LeaseID); err == nil {
				actions = append(actions, "lease_revoked")
			}
		}
		if req.SecretPath != "" && h.rotation != nil {
			policy, err := h.rotation.GetPolicy(c.Request.Context(), req.SecretPath)
			if err == nil && policy != nil {
				if err := h.rotation.RotatePath(c.Request.Context(), policy); err == nil {
					actions = append(actions, "kv_rotated")
				}
			}
		}
	}
	if h.exposureWebhook != nil {
		_ = h.exposureWebhook.Send(c.Request.Context(), notify.Event{
			Event:    "exposure.reported",
			Path:     req.SecretPath,
			LeaseID:  req.LeaseID,
			Severity: req.Severity,
			Detector: req.Detector,
			Details: map[string]any{
				"fingerprint": req.Fingerprint,
				"actions":     actions,
			},
		})
	}
	c.JSON(http.StatusOK, gin.H{
		"reported": true,
		"actions":  actions,
	})
}

// RotateMasterKey handles POST /sys/rotate-master-key.
func (h *SysHandler) RotateMasterKey(c *gin.Context) {
	if h.masterKey == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "master key rotation not configured"))
		return
	}
	var req dto.RotateMasterKeyRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	if strings.TrimSpace(req.NewKey) == "" {
		_ = c.Error(common.New(common.ErrCodeValidation, "new_key is required"))
		return
	}
	result, err := h.masterKey.Rotate(c.Request.Context(), service.RotateRequest{NewKeyBase64: req.NewKey})
	if err != nil {
		_ = c.Error(err)
		return
	}
	c.JSON(http.StatusOK, dto.RotateMasterKeyResponse{KeyVersion: int(result.KeyVersion)})
}

// Seal handles POST /sys/seal.
func (h *SysHandler) Seal(c *gin.Context) {
	if h.seal == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "seal not configured"))
		return
	}
	h.seal.Seal()
	c.JSON(http.StatusOK, gin.H{"sealed": true})
}

// Unseal handles POST /sys/unseal (full key or Shamir share).
func (h *SysHandler) Unseal(c *gin.Context) {
	if h.seal == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "seal not configured"))
		return
	}
	// W75-04: optional source IP allowlist for unseal.
	if len(h.unsealAllowNets) > 0 {
		if !netutil.IPAllowed(c.ClientIP(), h.unsealAllowNets) {
			_ = c.Error(common.New(common.ErrCodeForbidden, "unseal not allowed from this source"))
			return
		}
	}
	var req dto.UnsealRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	// W50-28: progressive backoff between failed unseal attempts.
	if ra, ok := h.seal.(interface{ UnsealRetryAfter() time.Duration }); ok {
		if wait := ra.UnsealRetryAfter(); wait > 0 {
			c.Header("Retry-After", fmt.Sprintf("%d", int(wait.Seconds())+1))
			_ = c.Error(common.New(common.ErrCodeForbidden, "unseal rate limited; retry later"))
			return
		}
	}
	// Multi-share path
	if req.Share != "" {
		raw, err := base64.StdEncoding.DecodeString(req.Share)
		if err != nil {
			_ = c.Error(common.Wrap(common.ErrCodeValidation, "invalid base64 unseal share", err))
			return
		}
		ms, ok := h.seal.(MultiShareUnsealer)
		if !ok {
			_ = c.Error(common.New(common.ErrCodeValidation, "multi-share unseal not supported"))
			return
		}
		unsealed, have, need, errMsg := ms.SubmitShare(raw)
		if errMsg != "" && !unsealed {
			_ = c.Error(common.New(common.ErrCodeValidation, errMsg))
			return
		}
		c.JSON(http.StatusOK, dto.UnsealResponse{Sealed: !unsealed, Progress: have, Threshold: need})
		return
	}
	if req.Key == "" {
		_ = c.Error(common.New(common.ErrCodeValidation, "key or share is required"))
		return
	}
	raw, err := base64.StdEncoding.DecodeString(req.Key)
	if err != nil {
		_ = c.Error(common.Wrap(common.ErrCodeValidation, "invalid base64 unseal key", err))
		return
	}
	if !h.seal.Unseal(raw) {
		_ = c.Error(common.New(common.ErrCodeValidation, "invalid unseal key"))
		return
	}
	c.JSON(http.StatusOK, dto.UnsealResponse{Sealed: false})
}

// GenerateUnsealShares splits a provided unseal key into Shamir shares (admin).
func (h *SysHandler) GenerateUnsealShares(c *gin.Context) {
	var req dto.SplitUnsealRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	if req.Shares < 2 || req.Threshold < 1 || req.Threshold > req.Shares {
		_ = c.Error(common.New(common.ErrCodeValidation, "need 1 <= threshold <= shares and shares >= 2"))
		return
	}
	raw, err := base64.StdEncoding.DecodeString(req.Key)
	if err != nil || len(raw) == 0 {
		_ = c.Error(common.New(common.ErrCodeValidation, "key must be base64-encoded unseal secret"))
		return
	}
	parts, err := shamir.Split(raw, req.Shares, req.Threshold)
	if err != nil {
		_ = c.Error(common.Wrap(common.ErrCodeValidation, "split unseal key", err))
		return
	}
	out := make([]string, len(parts))
	for i, p := range parts {
		out[i] = base64.StdEncoding.EncodeToString(p)
	}
	c.JSON(http.StatusOK, dto.SplitUnsealResponse{Shares: out, Threshold: req.Threshold})
}

func (h *SysHandler) requireRaftLeader(c *gin.Context) bool {
	if h.raft == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "raft membership not configured"))
		return false
	}
	if !h.raft.IsLeader() {
		_ = c.Error(common.New(common.ErrCodeValidation, "raft membership changes must be proposed on the leader"))
		return false
	}
	return true
}

// RaftAddNode handles POST /sys/raft/add-node.
func (h *SysHandler) RaftAddNode(c *gin.Context) {
	if !h.requireRaftLeader(c) {
		return
	}
	var req dto.RaftAddNodeRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	if err := h.raft.AddNode(c.Request.Context(), req.NodeID, req.Address); err != nil {
		_ = c.Error(common.Wrap(common.ErrCodeInternal, "add raft node", err))
		return
	}
	c.JSON(http.StatusOK, gin.H{"added": true, "node_id": req.NodeID})
}

// RaftRemoveNode handles POST /sys/raft/remove-node.
func (h *SysHandler) RaftRemoveNode(c *gin.Context) {
	if !h.requireRaftLeader(c) {
		return
	}
	var req dto.RaftRemoveNodeRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}
	if err := h.raft.RemoveNode(c.Request.Context(), req.NodeID); err != nil {
		_ = c.Error(common.Wrap(common.ErrCodeInternal, "remove raft node", err))
		return
	}
	c.JSON(http.StatusOK, gin.H{"removed": true, "node_id": req.NodeID})
}

// RunRotation handles POST /sys/rotation/run.
func (h *SysHandler) RunRotation(c *gin.Context) {
	if h.orchestration == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "rotation orchestration not configured"))
		return
	}
	var req dto.RotationRunRequest
	_ = c.ShouldBindJSON(&req)

	dbGrace := 24 * time.Hour
	if req.DBGrace != "" {
		parsed, err := utils.ParseTTL(req.DBGrace)
		if err != nil {
			_ = c.Error(common.Wrap(common.ErrCodeValidation, "invalid db_grace", err))
			return
		}
		dbGrace = parsed
	}
	sshGrace := dbGrace
	if req.SSHGrace != "" {
		parsed, err := utils.ParseTTL(req.SSHGrace)
		if err != nil {
			_ = c.Error(common.Wrap(common.ErrCodeValidation, "invalid ssh_grace", err))
			return
		}
		sshGrace = parsed
	}
	pkiGrace := 72 * time.Hour
	if req.PKIGrace != "" {
		parsed, err := utils.ParseTTL(req.PKIGrace)
		if err != nil {
			_ = c.Error(common.Wrap(common.ErrCodeValidation, "invalid pki_grace", err))
			return
		}
		pkiGrace = parsed
	}

	result, err := h.orchestration.Run(c.Request.Context(), dbGrace, sshGrace, pkiGrace)
	if err != nil {
		_ = c.Error(err)
		return
	}
	c.JSON(http.StatusOK, result)
}

// IssueListenerTLS issues a listener certificate and optionally writes PEM files.
func (h *SysHandler) IssueListenerTLS(c *gin.Context) {
	if h.pki == nil {
		_ = c.Error(common.New(common.ErrCodeInternal, "pki not configured"))
		return
	}
	var req dto.IssueListenerTLSRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		_ = c.Error(err)
		return
	}

	ttl := req.TTL
	if ttl == "" {
		ttl = "8760h"
	}
	result, err := h.pki.IssueCertificate(c.Request.Context(), pkiengine.IssueRequest{
		Role:       req.Role,
		CommonName: req.CommonName,
		DNSNames:   req.DNSNames,
		TTL:        ttl,
		AutoRenew:  true,
	})
	if err != nil {
		_ = c.Error(err)
		return
	}

	resp := dto.IssueListenerTLSResponse{
		CertPEM:       result.CertPEM,
		PrivateKeyPEM: result.PrivateKeyPEM,
		Serial:        result.Serial,
		ExpiresAt:     result.ExpiresAt.Format(time.RFC3339),
	}
	if req.CertFile != "" && req.KeyFile != "" {
		if err := writeListenerPEM(req.CertFile, req.KeyFile, result.CertPEM, result.PrivateKeyPEM); err != nil {
			_ = c.Error(common.Wrap(common.ErrCodeInternal, "write listener tls files", err))
			return
		}
		resp.CertFile = req.CertFile
		resp.KeyFile = req.KeyFile
	}
	c.JSON(http.StatusOK, resp)
}

func writeListenerPEM(certFile, keyFile, certPEM, keyPEM string) error {
	for _, path := range []string{certFile, keyFile} {
		if err := os.MkdirAll(filepath.Dir(path), 0o700); err != nil {
			return err
		}
	}
	if err := os.WriteFile(certFile, []byte(certPEM), 0o600); err != nil {
		return err
	}
	return os.WriteFile(keyFile, []byte(keyPEM), 0o600)
}